fix wrong package name for storm trident
